requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
argcomplete==1.12.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-APScheduler==1.11.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pika==1.1.0
Pillow==7.2.0
pyserial==3.4
python-dateutil==2.8.1
python-escpos==2.2.0
pytz==2020.1
pyusb==1.0.2
PyYAML==5.3.1
pyzk==0.9
qrcode==6.1
requests==2.24.0
six==1.15.0
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
